Remove PendingOffersTab folder, move logic to PendingOffersTable
