--- src/pages/OffersPage/components/PendingOffersTab/PendingOffersTab.tsx
@@ -1,10 +1,0 @@
-import { PendingOfferTable } from '../PendingOffersTable'
-import { usePendingOfferTab } from './hooks'
-
-const PendingOffersTab = () => {
-  const { offers, loading, sortViewParams } = usePendingOfferTab()
-
-  return <PendingOfferTable data={offers} loading={loading} sortViewParams={sortViewParams} />
-}
-
-export default PendingOffersTab
