--- src/pages/OffersPage/components/PendingOffersTab/index.ts
@@ -1,1 +1,0 @@
-export { default } from './PendingOffersTab'
