+++ src/pages/OffersPage/components/PendingOffersTable/constants.ts
